- Fixed Database connection

+++ src/main/java/com/project23/app/repository/BusinessObjectRepository.java
@@ -3,5 +3,5 @@ package com.project23.app.repository;
 import com.project23.app.pojo.BusinessObject;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusinessObjectRepository extends JpaRepository<BusinessObject, Integer> {
+public interface BusinessObjectRepository extends JpaRepository<BusinessObject, Long> {
 }
