+++ src/main/java/com/project23/app/repository/LabelRepository.java
@@ -3,5 +3,5 @@ package com.project23.app.repository;
 import com.project23.app.pojo.Label;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LabelRepository extends JpaRepository<Label, Integer> {
+public interface LabelRepository extends JpaRepository<Label, Long> {
 }
