+++ src/main/java/com/project23/app/repository/SourceSystemRepository.java
@@ -3,5 +3,5 @@ package com.project23.app.repository;
 import com.project23.app.pojo.SourceSystem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SourceSystemRepository extends JpaRepository<SourceSystem, Integer> {
+public interface SourceSystemRepository extends JpaRepository<SourceSystem, Long> {
 }
